final commit

+++ src/org/processmining/specpp/evaluation/heuristics/EventuallyFollowsTreeTraversalHeuristic.java
@@ -157,7 +157,7 @@ public class EventuallyFollowsTreeTraversalHeuristic implements HeuristicStrateg
                 .flatMap(i -> node.getPlace().postset().streamIndices().map(j -> eventuallyFollows[i][j]))
                 .sum();
 
-        double score = alpha * ((double) sum / (node.getPlace().preset().size() * node.getPlace().postset().size()) / maxEF) + (1-alpha) * (1 - ((double) node.getPlace().size() / maxSize));
+        double score = alpha * (sum / (node.getPlace().preset().size() * node.getPlace().postset().size()) / maxEF) + (1-alpha) * (1 - ((double) node.getPlace().size() / maxSize));
         return new TreeNodeScore(score);
     }
